Fix compile errors

+++ nekoyume/Assets/_Scripts/UI/Module/AvatarStats.cs
@@ -32,7 +32,7 @@ namespace Nekoyume.UI.Module
                     }
 
                     var (statType, value) = enumerator.Current;
-                    statView.Show(statType, (int)value, 0);
+                    statView.Show(statType, value, 0);
                 }
             }
         }
